WIP to be reviewed: TUA-468 API test implementation

+++ src/main/java/org/ssu/edu/teachua/api/clients/ClubClient.java
@@ -5,6 +5,8 @@ import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import org.ssu.edu.teachua.api.models.club.ClubRequest;
 
+import java.math.BigInteger;
+
 public class ClubClient extends BaseClient {
 
     public ClubClient(String url, ContentType contentType, String accessToken) {
@@ -22,7 +24,7 @@ public class ClubClient extends BaseClient {
     }
 
     @Step("Create a request for updating an existing club entity using 'PUT' method and club id {id} as parameter. Enter values in 'Body': {request}")
-    public Response editClub(ClubRequest request, int id) {
+    public Response editClub(ClubRequest request, BigInteger id) {
         return prepareRequest()
                 .body(request)
                 .when()
@@ -30,14 +32,14 @@ public class ClubClient extends BaseClient {
     }
 
     @Step("Create a request for viewing an existing club entity using 'GET' method and club id {id} as parameter")
-    public Response viewClub(int id) {
+    public Response viewClub(BigInteger id) {
         return prepareRequest()
                 .when()
                 .get(path + "/" + id);
     }
 
     @Step("Create a request for deleting an existing club entity using 'DELETE' method and club id {id} as parameter")
-    public Response deleteClub(int id) {
+    public Response deleteClub(BigInteger id) {
         return prepareRequest()
                 .when()
                 .delete(path + "/" + id);
